Fix false non-private variables detection.
To properly set the private status of a variable in an upper context it is not enough to have only is_private flag.
There should be a flag indicating there was no read before writing to the variable.
Assume an AddrInfo object could not exist without any access. Otherwise, rewrite the code to set the is_init flag based on a read operation.